Lock down css-loader to version 2 to fix Ⓜ in class names
Version 3+ of css-loader uses `cssesc` which does not maintain Ⓜ when generating classnames in outputted CSS :-(

Yarn does give a warning that we are not providing the right version of webpack for css-loader because version 2 wants webpack version 4, but it does work.

In the future, if I need a newer version of css-loader, I'll probably have to fork css-loader.