Cleanup the webpack stats the best I could for v5
V5 of webpack changed the stats to be organized differently. I tried to get the stats to look the same as V4 but still a bit verbose. This is the closes we can do for now.

There may be improvements in the future so I should take another run at this in a year or so.