Support CSS properties in browsers which don't support them like IE11
Without a reference of the properties, there was no way for `postcss-preset-env` to render the values for browsers that don't support them.

Now this will look for a `variables.CSS` file in both the JS directory and the theme's pcss and load the values if available.

The files must have a .CSS extension or they won't load.